# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Fleet EDR API
  version: "0.1.0"
  description: |
    HTTP API for Fleet EDR. Three audience tiers:

    - **Agents** (macOS endpoints) authenticate with a per-host bearer
      token obtained at enrollment. They call `POST /api/enroll`,
      `POST /api/events`, and the host-scoped `/commands` endpoints.
    - **Browsers** (admin UI) authenticate with an `edr_session` cookie
      (HttpOnly, Secure, SameSite=Lax) plus an `X-CSRF-Token` header on
      unsafe methods. They call the session-protected endpoints.
    - **Operators** hit `/livez`, `/readyz`, `/health` with no auth.

    For prose context see
    [docs/api.md](https://github.com/getvictor/fleet-edr/blob/main/docs/api.md).
    For a human-readable host/agent install guide see
    [docs/install-agent-manual.md](https://github.com/getvictor/fleet-edr/blob/main/docs/install-agent-manual.md).
  license:
    name: Fleet EDR License
    url: https://github.com/getvictor/fleet-edr/blob/main/LICENSE
servers:
  - url: https://edr.example.com
    description: Your EDR server
...
paths:
...
  /api/session:
    post:
      tags: [Session]
      summary: Log in with email + password
      description: |
        Validates credentials, creates a session row, sets `edr_session`
        cookie, returns the user + CSRF token. Rate-limited per-IP at
        `EDR_LOGIN_RATE_PER_MIN` (default 6).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          headers:
            Set-Cookie:
              description: Session cookie (HttpOnly, Secure, SameSite=Lax)
              schema:
                type: string
                example: edr_session=<opaque>; Path=/; HttpOnly; Secure; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Malformed request
        "401":
          description: Wrong email or password
        "429":
          description: Rate limited
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema: {type: integer}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags: [Session]
      summary: Get the current session
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Not logged in
    delete:
      tags: [Session]
      summary: Log out
      description: Clears the session row and expires the cookie.
      security:
        - SessionCookie: []
          CSRF: []
      responses:
        "204":
          description: Logged out
...
components:
  securitySchemes:
    HostToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Per-host bearer token issued by `POST /api/enroll`. Pass as
        `Authorization: Bearer <host_token>`.
    SessionCookie:
      type: apiKey
      in: cookie
      name: edr_session
      description: |
        HttpOnly cookie set by `POST /api/session`. The browser
        attaches it automatically.
    CSRF:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Required on unsafe methods (`POST`, `PUT`, `DELETE`) for
        session-cookie-authenticated endpoints. Value comes from the
        login response's `csrf_token` field.
...
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password}
    LoginResponse:
      type: object
      required: [user, csrf_token]
      properties:
        user:
          type: object
          required: [id, email]
          properties:
            id: {type: integer, format: int64}
            email: {type: string, format: email}
        csrf_token: {type: string}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: {type: string}
...